rename:common state 관련 파일들 케밥케이스로 통일
