+++ app/feature/curation/components/CurationDetail/organisms/CurationDetailPlaceCard.tsx
@@ -3,11 +3,11 @@ import { CurationPlaceProps } from "@/feature/curation/type";
 import LinkLayout from "@/common/components/layout/Link/LinkLayout";
 import PlaceInfoCardBottom from "@/feature/place/components/PlaceInfo/organisms/PlaceInfoCardBottom";
 import { useSetRecoilState } from "recoil";
-import { myCurationModalInfoSelector } from "@/common/state/myCurationModal";
+import { myCurationModalInfoSelector } from "@/common/state/my-curation-modal";
 import { toastInfoSelector } from "@/common/state/toast";
 import { validateLoggedIn } from "@/common/utils/validate/validate-logged-in";
 import useCurationSpaceDelete from "@/feature/curation/queries/useCurationSpaceDelete";
-import { queryFetchingSelector } from "@/common/state/queryFetching";
+import { queryFetchingSelector } from "@/common/state/query-fetching";
 import ScrapFillIcon from "@/common/assets/icons/scrap/ScrapFillIcon";
 import ScrapLineIcon from "@/common/assets/icons/scrap/ScrapLineIcon";
 import { useInView } from "react-intersection-observer";
