+++ app/feature/place/components/PlaceDetail/molecules/PlaceScrapIcon.tsx
@@ -2,7 +2,7 @@
 
 import ScrapFillIcon from "@/common/assets/icons/scrap/ScrapFillIcon";
 import ScrapLineIcon from "@/common/assets/icons/scrap/ScrapLineIcon";
-import { myCurationModalInfoSelector } from "@/common/state/myCurationModal";
+import { myCurationModalInfoSelector } from "@/common/state/my-curation-modal";
 import { toastInfoSelector } from "@/common/state/toast";
 import { validateLoggedIn } from "@/common/utils/validate/validate-logged-in";
 import useGetPlaceScrapState from "@/feature/place/queries/useGetPlaceScrapState";
